requirements.txt
# Core framework
streamlit>=1.40.0
anthropic>=0.86.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data / financial
yfinance>=0.2.50
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# RAG / embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ML / classification
scikit-learn>=1.5.0

# Sentiment
vaderSentiment>=3.3.2

# Satellite (optional, if GEE available)
earthengine-api>=0.1.380

# Visualization
plotly>=5.20.0
matplotlib>=3.8.0

# Utilities
httpx>=0.27.0
python-dateutil>=2.9.0
pyyaml>=6.0
